+++ src/xyz/oihana/schema/constants/traits/business/documents/BusinessDocumentLineTrait.php
@@ -15,10 +15,12 @@ trait BusinessDocumentLineTrait
     const string ADJUSTMENTS         = 'adjustments' ;
     const string COLOR               = 'color' ;
     const string FREE_REASON         = 'freeReason' ;
+    const string INCLUDED_IN_TOTAL   = 'includedInTotal' ;
     const string ITEM                = 'item' ;
     const string POSITION            = 'position' ;
     const string PRICE               = 'price' ;
     const string QUANTITY            = 'quantity' ;
+    const string SECTION             = 'section' ;
     const string SUBTOTAL            = 'subtotal' ;
     const string TAXES               = 'taxes' ;
     const string TECHNICAL_NOTE      = 'technicalNote' ;